Adding empty translations, fixing a few
There were several empty translations at the beginning of the file. Also, there were some incomplete or wrong.
It is not clear whether the messages are intended to address the user in a formal or informal way (similar to the difference between you and thou). This could be included in the README for clarity.